+++ src/main/java/com/elara/app/unit_of_measure_service/service/implementation/UomStatusServiceImp.java
@@ -1,15 +1,13 @@
-package com.elara.app.unit_of_measure_service.service.imp;
+package com.elara.app.unit_of_measure_service.service.implementation;
 
-import com.elara.app.unit_of_measure_service.dto.request.UomRequest;
-import com.elara.app.unit_of_measure_service.dto.response.UomResponse;
-import com.elara.app.unit_of_measure_service.dto.update.UomUpdate;
+import com.elara.app.unit_of_measure_service.dto.request.UomStatusRequest;
+import com.elara.app.unit_of_measure_service.dto.response.UomStatusResponse;
+import com.elara.app.unit_of_measure_service.dto.update.UomStatusUpdate;
 import com.elara.app.unit_of_measure_service.exceptions.ResourceConflictException;
 import com.elara.app.unit_of_measure_service.exceptions.ResourceNotFoundException;
-import com.elara.app.unit_of_measure_service.mapper.UomMapper;
-import com.elara.app.unit_of_measure_service.model.Uom;
+import com.elara.app.unit_of_measure_service.mapper.UomStatusMapper;
 import com.elara.app.unit_of_measure_service.model.UomStatus;
-import com.elara.app.unit_of_measure_service.repository.UomRepository;
-import com.elara.app.unit_of_measure_service.service.interfaces.UomService;
+import com.elara.app.unit_of_measure_service.repository.UomStatusRepository;
 import com.elara.app.unit_of_measure_service.service.interfaces.UomStatusService;
 import com.elara.app.unit_of_measure_service.utils.MessageService;
 import jakarta.transaction.Transactional;
@@ -25,18 +23,17 @@ import java.util.Optional;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class UomServiceImp implements UomService {
+public class UomStatusServiceImp implements UomStatusService {
 
-    private static final String ENTITY_NAME = "Uom";
-    private static final String NOMENCLATURE = "Uom-service";
-    private final UomRepository repository;
-    private final UomMapper mapper;
+    private static final String ENTITY_NAME = "UomStatus";
+    private static final String NOMENCLATURE = ENTITY_NAME + "-service";
+    private final UomStatusRepository repository;
+    private final UomStatusMapper mapper;
     private final MessageService messageService;
-    private final UomStatusService statusService;
 
     @Override
     @Transactional
-    public UomResponse save(UomRequest request) {
+    public UomStatusResponse save(UomStatusRequest request) {
         final String methodNomenclature = NOMENCLATURE + "-save";
         log.info("[{}] {} record to save: {}", methodNomenclature, ENTITY_NAME, request);
         try {
@@ -45,13 +42,11 @@ public class UomServiceImp implements UomService {
                 log.warn("[{}] {}", methodNomenclature, alreadyExistsMsg);
                 throw new ResourceConflictException(new Object[]{alreadyExistsMsg});
             }
-            Uom entity = mapper.toEntity(request);
-            UomStatus status = statusService.findByIdService(request.uomStatusId());
-            entity.setUomStatus(status);
-            Uom saved = repository.save(entity);
+            UomStatus entity = mapper.toEntity(request);
+            UomStatus saved = repository.save(entity);
             log.info("[{}] {} record created with id: {}.", methodNomenclature, ENTITY_NAME, saved.getId());
             return mapper.toResponse(saved);
-        } catch (ResourceConflictException | ResourceNotFoundException e) {
+        } catch (ResourceConflictException e) {
             String saveErrorMsg = messageService.getMessage("crud.save.error", ENTITY_NAME);
             log.warn("[{}] {}", methodNomenclature, saveErrorMsg);
             throw e;
@@ -60,11 +55,11 @@ public class UomServiceImp implements UomService {
 
     @Override
     @Transactional
-    public UomResponse update(Long id, UomUpdate request) {
+    public UomStatusResponse update(Long id, UomStatusUpdate request) {
         final String methodNomenclature = NOMENCLATURE + "-update";
         log.info("[{}] Update {} record with id: {} and request: {}", methodNomenclature, ENTITY_NAME, id, request);
         try {
-            Uom existing = repository.findById(id)
+            UomStatus existing = repository.findById(id)
                 .orElseThrow(() -> {
                     String notFoundMsg = messageService.getMessage("crud.not.found", ENTITY_NAME, "id", id);
                     log.warn("[{}] {}", methodNomenclature, notFoundMsg);
@@ -75,9 +70,7 @@ public class UomServiceImp implements UomService {
                 log.warn("[{}] {}", methodNomenclature, alreadyExistsMsg);
                 throw new ResourceConflictException(new Object[]{alreadyExistsMsg});
             }
-            UomStatus status = existing.getUomStatus();
             mapper.updateEntityFromDto(existing, request);
-            existing.setUomStatus(status);
             log.info("[{}] {} record updated with data: {}", methodNomenclature, ENTITY_NAME, existing);
             return mapper.toResponse(existing);
         } catch (ResourceNotFoundException | ResourceConflictException e) {
@@ -88,6 +81,7 @@ public class UomServiceImp implements UomService {
     }
 
     @Override
+    @Transactional
     public void deleteById(Long id) {
         final String methodNomenclature = NOMENCLATURE + "-deleteById";
         log.debug("[{}] Delete {} record with id: {}", methodNomenclature, ENTITY_NAME, id);
@@ -107,11 +101,11 @@ public class UomServiceImp implements UomService {
     }
 
     @Override
-    public UomResponse findById(Long id) {
+    public UomStatusResponse findById(Long id) {
         final String methodNomenclature = NOMENCLATURE + "-findById";
         log.info("[{}] Fetch {} record with id: {}", methodNomenclature, ENTITY_NAME, id);
         try {
-            Optional<Uom> entity = repository.findById(id);
+            Optional<UomStatus> entity = repository.findById(id);
             if (entity.isEmpty()) {
                 String msg = messageService.getMessage("crud.not.found", ENTITY_NAME, "id", id);
                 log.warn("[{}] {}", methodNomenclature, msg);
@@ -126,62 +120,67 @@ public class UomServiceImp implements UomService {
         }
     }
 
+    @Transactional
+    public UomStatus findByIdService(Long id) {
+        final String methodNomenclature = NOMENCLATURE + "-findByIdService";
+        log.info("[{}] Fetch {} record with id: {}", methodNomenclature, ENTITY_NAME, id);
+        UomStatus entity = repository.findById(id)
+            .orElseThrow(() -> {
+                String msg = messageService.getMessage("crud.not.found", ENTITY_NAME, "id", id);
+                log.warn("[{}] {}", methodNomenclature, msg);
+                return new ResourceNotFoundException(new Object[]{ENTITY_NAME, "id", id.toString()});
+            });
+        log.debug("[{}] Fetched {} record with id: {}: {}", methodNomenclature, ENTITY_NAME, id, entity);
+        return entity;
+    }
+
     @Override
-    public Page<UomResponse> findAll(Pageable pageable) {
+    public Page<UomStatusResponse> findAll(Pageable pageable) {
         final String methodNomenclature = NOMENCLATURE + "-findAll";
         log.debug("[{}] Fetch all {} records.", methodNomenclature, ENTITY_NAME);
-        Page<UomResponse> page = repository.findAll(pageable).map(mapper::toResponse);
+        Page<UomStatusResponse> page = repository.findAll(pageable).map(mapper::toResponse);
         log.debug("[{}] Fetched {} {} records.", methodNomenclature, page.getNumberOfElements(), ENTITY_NAME);
         return page;
     }
 
     @Override
-    public Page<UomResponse> findAllByName(String name, Pageable pageable) {
+    public Page<UomStatusResponse> findAllByName(String name, Pageable pageable) {
         final String methodNomenclature = NOMENCLATURE + "-findAllByName";
         log.debug("[{}] Fetch all {} records that contain in their name: '{}'", methodNomenclature, ENTITY_NAME, name);
-        Page<UomResponse> page = repository.findAllByNameContainingIgnoreCase(name, pageable).map(mapper::toResponse);
+        Page<UomStatusResponse> page = repository.findAllByNameContainingIgnoreCase(name, pageable).map(mapper::toResponse);
         log.info("[{}] Fetched {} {} entities with name like '{}'.", methodNomenclature, page.getNumberOfElements(), ENTITY_NAME, name);
         return page;
     }
 
     @Override
-    public Page<UomResponse> findAllByUomStatusId(Long uomStatusId, Pageable pageable) {
-        final String methodNomenclature = NOMENCLATURE + "-findAllByUomStatusId";
-        log.debug("[{}] Fetch all {} records with status id: '{}'", methodNomenclature, ENTITY_NAME, uomStatusId);
-        Page<UomResponse> page = repository.findAllByUomStatusId(uomStatusId, pageable).map(mapper::toResponse);
-        log.info("[{}] Fetched {} {} records with status id: '{}'", methodNomenclature, page.getNumberOfElements(), ENTITY_NAME, uomStatusId);
+    public Page<UomStatusResponse> findAllByIsUsable(Boolean isUsable, Pageable pageable) {
+        final String methodNomenclature = NOMENCLATURE + "-findAllByIsUsable";
+        log.info("[{}] Fetch all {} records with isUsable: {}.", methodNomenclature, ENTITY_NAME, isUsable);
+        Page<UomStatusResponse> page = repository.findAllByIsUsable(isUsable, pageable).map(mapper::toResponse);
+        log.info("[{}] Fetched {} {} records with isUsable: {}.", methodNomenclature, page.getNumberOfElements(), ENTITY_NAME, isUsable);
         return page;
     }
 
-    @Override
     public Boolean isNameTaken(String name) {
         final String methodNomenclature = NOMENCLATURE + "-isNameTaken";
         log.info("[{}] Check if name '{}' is taken.", methodNomenclature, name);
-        Boolean exists = repository.existsByNameIgnoreCase(name);
+        Boolean exists = repository.existsByName(name);
         log.info("[{}] Name '{}' {} taken.", methodNomenclature, name, exists ? "is" : "is not");
         return exists;
     }
 
     @Override
     @Transactional
-    public void changeStatus(Long id, Long uomStatusId) {
+    public void changeStatus(Long id, Boolean isUsable) {
         final String methodNomenclature = NOMENCLATURE + "-changeStatus";
-        log.debug("[{}] Change status id of {} record with id: {} to: {}", methodNomenclature, ENTITY_NAME, id, uomStatusId);
-        try {
-            Uom existing = repository.findById(id)
-                .orElseThrow(() -> {
-                    String msg = messageService.getMessage("crud.not.found", ENTITY_NAME, "id", id);
-                    log.warn("[{}] {}", methodNomenclature, msg);
-                    return new ResourceNotFoundException(new Object[]{ENTITY_NAME, "id", id});
-                });
-            UomStatus newStatus = statusService.findByIdService(uomStatusId);
-            existing.setUomStatus(newStatus);
-            log.debug("[{}] Changed status id of {} record with id: {} to: {}", methodNomenclature, ENTITY_NAME, id, newStatus.getId());
-        } catch (ResourceNotFoundException e) {
-            String updateErrorMsg = messageService.getMessage("crud.update.error", ENTITY_NAME);
-            log.warn("[{}] {}", methodNomenclature, updateErrorMsg);
-            throw e;
-        }
+        log.info("[{}] Change status of {} record with id: {} to: {}", methodNomenclature, ENTITY_NAME, id, isUsable);
+        UomStatus existing = repository.findById(id)
+            .orElseThrow(() -> {
+                String msg = messageService.getMessage("crud.not.found", ENTITY_NAME, "id", id);
+                log.warn("[{}] {}", methodNomenclature, msg);
+                return new ResourceNotFoundException(new Object[]{ENTITY_NAME, "id", id});
+            });
+        existing.setIsUsable(isUsable);
+        log.info("[{}] Changed status of {} record with id: {} to: {}", methodNomenclature, ENTITY_NAME, id, isUsable);
     }
-
 }
